chore: rename multiple select dropdown
